Hunger Calculation method per-species
Update Hunger.md

Create Hunger.md

clean up

AlertSOs for hunger are handled by their status effects now

AAAAAAAAAAAA

aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa

Don't hardcode stomach capacity

fix NRE with advanced health scanners

make sure methods are registered in assets
