+++ UnityProject/Assets/Prefabs/Player/Resources/BodyParts/Moth/Moth.asset.bak
@@ -63,7 +63,7 @@ MonoBehaviour:
     version: 2
     RefIds:
     - rid: 5920833068371804161
-      type: {class: ReagentPoolSystem, ns: HealthV2.Living.PolymorphicSystems, asm: Assets}
+      type: {class: ReagentPoolSystem, ns: US13.HealthV2.Living.PolymorphicSystems, asm: Assets}
       data:
         Base: {fileID: 0}
         BloodPool:
@@ -84,7 +84,7 @@ MonoBehaviour:
           reagentKeys: []
         ReservoirReturnMultiplier: 0.15
     - rid: 5920833068371804162
-      type: {class: ChemicalMetabolismSystem, ns: HealthV2.Living.PolymorphicSystems,
+      type: {class: ChemicalMetabolismSystem, ns: US13.HealthV2.Living.PolymorphicSystems,
         asm: Assets}
       data:
         Base: {fileID: 0}
@@ -95,7 +95,7 @@ MonoBehaviour:
         ALLMetabolismReactions: []
         MetabolismComponents: []
     - rid: 5920833068371804164
-      type: {class: HungerSystem, ns: HealthV2.Living.PolymorphicSystems, asm: Assets}
+      type: {class: HungerSystem, ns: US13.HealthV2.Living.PolymorphicSystems, asm: Assets}
       data:
         Base: {fileID: 0}
         BodyParts: []
@@ -113,7 +113,7 @@ MonoBehaviour:
           FatStatusEffect: {fileID: 11400000, guid: 8a92c094201a758498a79523323565eb,
             type: 2}
     - rid: 5920833068371804166
-      type: {class: NaturalChemicalReleaseSystem, ns: HealthV2.Living.PolymorphicSystems,
+      type: {class: NaturalChemicalReleaseSystem, ns: US13.HealthV2.Living.PolymorphicSystems,
         asm: Assets}
       data:
         Base: {fileID: 0}
