+++ UnityProject/Assets/Resources/ScriptableObjectsSingletons/AlertSOs.asset
@@ -29,11 +29,14 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 3803e6aed03ceaf4190e5fdbd1e27dbb, type: 2}
   - {fileID: 11400000, guid: 756fa4a7b0404794a9db193e0514e2b5, type: 2}
   - {fileID: 11400000, guid: 37880884910e549479c706e04e015105, type: 2}
+  - {fileID: 11400000, guid: 736bf1c84a056584c99884337fe16d86, type: 2}
   - {fileID: 11400000, guid: f4fac805011839b41b20e89a0937dd5d, type: 2}
   - {fileID: 11400000, guid: 71afda3df614b0c4f80411284b9168c7, type: 2}
   - {fileID: 11400000, guid: 2185c2eec20ef9444acb8b817bac1d13, type: 2}
   - {fileID: 11400000, guid: aa9f807a62b767242a56678ed42f09ba, type: 2}
   - {fileID: 11400000, guid: 218c15d167c2a3843ba514b45c9663c9, type: 2}
+  - {fileID: 11400000, guid: 4ad76f5ab439fb04dad532963740aef2, type: 2}
+  - {fileID: 11400000, guid: bb846b67e8fe7174a9259cac094036c9, type: 2}
   - {fileID: 11400000, guid: a242a18320bcd0345b83fe02838d0460, type: 2}
   - {fileID: 11400000, guid: bd2e15d992ea0ee499a3f1e0244d5c06, type: 2}
   - {fileID: 11400000, guid: 799b82ff02b152f42b86a463a5a35548, type: 2}
@@ -41,5 +44,3 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 73f14bb661d06e14b927be321458c28c, type: 2}
   - {fileID: 11400000, guid: 2c812669f183287449afa3e620bfcef6, type: 2}
   - {fileID: 11400000, guid: 2cacc34a5e9de974f9f8f4274ffaf54f, type: 2}
-  - {fileID: 11400000, guid: bb846b67e8fe7174a9259cac094036c9, type: 2}
-  - {fileID: 11400000, guid: 4ad76f5ab439fb04dad532963740aef2, type: 2}
