+++ UnityProject/Assets/ScriptableObjects/Statuses/Hunger/Full.asset
@@ -12,3 +12,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 57ba74246e3a41e2b0a713c2ba3d28d8, type: 3}
   m_Name: Full
   m_EditorClassIdentifier: Assets::Systems.StatusesAndEffects.Implementations.Hunger.BaseHungerStatusEffect
+  StatusAlert: {fileID: 11400000, guid: 9dcf0f63efe865449a8efde4713b5348, type: 2}
+  MessageOnEnteringStatus: Fat Fuck.
