+++ UnityProject/Assets/ScriptableObjects/Statuses/Hunger/Starving.asset
@@ -12,3 +12,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 57ba74246e3a41e2b0a713c2ba3d28d8, type: 3}
   m_Name: Starving
   m_EditorClassIdentifier: Assets::Systems.StatusesAndEffects.Implementations.Hunger.BaseHungerStatusEffect
+  StatusAlert: {fileID: 11400000, guid: 37880884910e549479c706e04e015105, type: 2}
+  MessageOnEnteringStatus: <color=yellow>You're straving..</color>
