+++ UnityProject/Assets/ScriptableObjects/Statuses/Hunger/WellFed.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a2cf78c09e3073c44bfbe12e9789cd05
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
