+++ UnityProject/Assets/ScriptableObjects/UIAlerts/Hunger/WellFed.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b1055f6767b8d254fb6634c7bbb32cda, type: 3}
+  m_Name: WellFed
+  m_EditorClassIdentifier: Assets::US13.UI.Core.Alerts.AlertSO
+  AssociatedSprite: {fileID: 11400000, guid: 4e5856ca9f6259d42bd183fc8a000f9a, type: 2}
+  HoverToolTip: You're not hungry, for now.
+  PlayerProtip: {fileID: 0}
+  DoNotShowIfPresent: []
