+++ UnityProject/Assets/ScriptableObjects/UIAlerts/Hunger/WellFed.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 736bf1c84a056584c99884337fe16d86
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
