+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/PolymorphicSystems/Hunger.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 53cae7c318dd49899c7210b3fb063f32
+timeCreated: 1774674393
