+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/PolymorphicSystems/Hunger/HungerCalculationMethods.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c9a02b8bd5364179b404365822c55456
+timeCreated: 1774674709
