+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/PolymorphicSystems/Hunger/HungerCalculationMethods/DefaultUnityStationHungerCalculation.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 177cf7d762df40da93f6cdc782186f86
+timeCreated: 1774674732
