+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/PolymorphicSystems/Hunger/HungerCalculationMethods/StomachFullness.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a61a88bb9e3744c1b600a12d3bf9378d
+timeCreated: 1774707629
