+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/PolymorphicSystems/Hunger/IHungerCalculation.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: af7fa5862df74068bea5c0c6136b3114
+timeCreated: 1774674451
