+++ UnityProject/Assets/Scripts/US13/Systems/Antagonists/Antags/Changeling/ChangelingAbility/Implementation/Regenerate.cs
@@ -10,6 +10,7 @@ using US13.HealthV2.Living;
 using US13.HealthV2.Living.BodyParts;
 using US13.HealthV2.Living.CirculatorySystem;
 using US13.HealthV2.Living.PolymorphicSystems;
+using US13.HealthV2.Living.PolymorphicSystems.Hunger;
 using US13.Managers.NetworkManagement;
 using US13.Player;
 using US13.UI.Core;
