+++ UnityProject/Assets/Textures/interface/UI/alerts/alerts_well_fed.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9d1de2691495b304e9843ac3335b626d, type: 3}
+  m_Name: alerts_well_fed
+  m_EditorClassIdentifier: RootSillys::SpriteDataSO
+  foreverID: DqRVhvoW9pk0Gvk3YjmO
+  Variance:
+  - Frames:
+    - sprite: {fileID: 21300000, guid: 8de9716f53ba2464e8709050ccf30228, type: 3}
+      secondDelay: 0.09
+  IsPalette: 0
+  DisplayName: Well Fed
