+++ UnityProject/Assets/Textures/interface/UI/alerts/alerts_well_fed.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4e5856ca9f6259d42bd183fc8a000f9a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
